AIService is finished out with basic functionality. Deck has had its documentation updated. GameAction has been modified with a constructor. GameController is being worked on to be implemented. GameService had functionality added to support GameCOntroller. GameState had functionality added to support GameController. Player had functionality modified and documentation added. Money is an int value. Matchmaking was removed, it is gameService

+++ api/src/main/java/com/pokerface/pokerapi/game/GameState.java
@@ -163,7 +163,7 @@ public class GameState {
 
     public boolean matchBet (int playerSeatID){
         Player player = players.get(playerSeatID);
-        double difference=minimumBet-pot.getBet(playerSeatID);
+        int difference=minimumBet-pot.getBet(playerSeatID);
         if (player.getCashOnHand()>=difference){
             pot.add(difference,playerSeatID);
             player.setCashOnHand(player.getCashOnHand()-difference);
@@ -174,7 +174,7 @@ public class GameState {
         return true;
     }
 
-    public boolean placeBet(int playerSeatID, double betAmount){
+    public boolean placeBet(int playerSeatID, int betAmount){
         Player player=players.get(playerSeatID);
         if (player.getCashOnHand()>=betAmount){
             pot.add(betAmount,playerSeatID);
